requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
Django==3.0.5
django-betterforms==1.2
django-betterforms-jpic==1.2.2
django-extra-views==0.13.0
django-reverse-admin==2.8.4
djangorestframework==3.11.0
idna==2.9
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
